+++ src/CrossValidation/Reports/AggregateReport.php
@@ -2,7 +2,7 @@
 
 namespace Rubix\ML\CrossValidation\Reports;
 
-use Rubix\ML\CrossValidation\Reports\Results\Report;
+use Rubix\ML\Report;
 use InvalidArgumentException;
 
 use function count;
@@ -84,7 +84,7 @@ class AggregateReport implements ReportGenerator
      *
      * @param (string|int|float)[] $predictions
      * @param (string|int|float)[] $labels
-     * @return \Rubix\ML\CrossValidation\Reports\Results\Report
+     * @return \Rubix\ML\Report
      */
     public function generate(array $predictions, array $labels) : Report
     {