+++ src/CrossValidation/Reports/ContingencyTable.php
@@ -4,7 +4,7 @@ namespace Rubix\ML\CrossValidation\Reports;
 
 use Rubix\ML\Estimator;
 use Rubix\ML\EstimatorType;
-use Rubix\ML\CrossValidation\Reports\Results\Report;
+use Rubix\ML\Report;
 use InvalidArgumentException;
 
 /**
@@ -38,7 +38,7 @@ class ContingencyTable implements ReportGenerator
      * @param (string|int)[] $predictions
      * @param (string|int)[] $labels
      * @throws \InvalidArgumentException
-     * @return \Rubix\ML\CrossValidation\Reports\Results\Report
+     * @return \Rubix\ML\Report
      */
     public function generate(array $predictions, array $labels) : Report
     {