+++ src/CrossValidation/Reports/MulticlassBreakdown.php
@@ -4,7 +4,7 @@ namespace Rubix\ML\CrossValidation\Reports;
 
 use Rubix\ML\Estimator;
 use Rubix\ML\EstimatorType;
-use Rubix\ML\CrossValidation\Reports\Results\Report;
+use Rubix\ML\Report;
 use InvalidArgumentException;
 
 use function count;
@@ -42,7 +42,7 @@ class MulticlassBreakdown implements ReportGenerator
      * @param (string|int)[] $predictions
      * @param (string|int)[] $labels
      * @throws \InvalidArgumentException
-     * @return \Rubix\ML\CrossValidation\Reports\Results\Report
+     * @return \Rubix\ML\Report
      */
     public function generate(array $predictions, array $labels) : Report
     {