+++ src/CrossValidation/Reports/ReportGenerator.php
@@ -2,7 +2,7 @@
 
 namespace Rubix\ML\CrossValidation\Reports;
 
-use Rubix\ML\CrossValidation\Reports\Results\Report;
+use Rubix\ML\Report;
 
 interface ReportGenerator
 {
@@ -18,7 +18,7 @@ interface ReportGenerator
      *
      * @param (string|int|float)[] $predictions
      * @param (string|int|float)[] $labels
-     * @return \Rubix\ML\CrossValidation\Reports\Results\Report
+     * @return \Rubix\ML\Report
      */
     public function generate(array $predictions, array $labels) : Report;
 }