+++ tests/Datasets/LabeledTest.php
@@ -2,6 +2,7 @@
 
 namespace Rubix\ML\Tests\Datasets;
 
+use Rubix\ML\Report;
 use Rubix\ML\DataType;
 use Rubix\ML\Encoding;
 use Rubix\ML\Datasets\Dataset;
@@ -215,6 +216,18 @@ class LabeledTest extends TestCase
         $expected = array_transpose(self::SAMPLES);
 
         $this->assertEquals($expected, $this->dataset->columns());
+    }
+
+    /**
+     * @test
+     */
+    public function transformColumn() : void
+    {
+        $dataset = $this->dataset->transformColumn(3, 'abs');
+
+        $expected = [4.0, 1.5, 2.6, 1.0, 2.9, 5.0];
+
+        $this->assertEquals($expected, $dataset->column(3));
     }
 
     /**
@@ -718,8 +731,6 @@ class LabeledTest extends TestCase
      */
     public function describe() : void
     {
-        $stats = $this->dataset->describe();
-
         $expected = [
             [
                 'type' => 'categorical',
@@ -760,7 +771,10 @@ class LabeledTest extends TestCase
             ],
         ];
 
-        $this->assertEquals($expected, $stats);
+        $results = $this->dataset->describe();
+
+        $this->assertInstanceOf(Report::class, $results);
+        $this->assertEquals($expected, $results->toArray());
     }
 
     /**
@@ -843,24 +857,14 @@ class LabeledTest extends TestCase
                     'median' => -1.25,
                     '75%' => -1.125,
                     'max' => -1.0,
-
                 ],
             ],
         ];
 
-        $this->assertEquals($expected, $this->dataset->describeByLabel());
-    }
+        $results = $this->dataset->describeByLabel();
 
-    /**
-     * @test
-     */
-    public function transformColumn() : void
-    {
-        $dataset = $this->dataset->transformColumn(3, 'abs');
-
-        $expected = [4.0, 1.5, 2.6, 1.0, 2.9, 5.0];
-
-        $this->assertEquals($expected, $dataset->column(3));
+        $this->assertInstanceOf(Report::class, $results);
+        $this->assertEquals($expected, $results->toArray());
     }
 
     /**
@@ -868,8 +872,6 @@ class LabeledTest extends TestCase
      */
     public function describeLabels() : void
     {
-        $desc = $this->dataset->describeLabels();
-
         $expected = [
             'type' => 'categorical',
             'num_categories' => 2,
@@ -879,7 +881,10 @@ class LabeledTest extends TestCase
             ],
         ];
 
-        $this->assertEquals($expected, $desc);
+        $results = $this->dataset->describeLabels();
+
+        $this->assertInstanceOf(Report::class, $results);
+        $this->assertEquals($expected, $results->toArray());
     }
 
     /**